Fix specs
